+++ .gitignore
@@ -1,2 +1,3 @@
 *.o
 vwm
+*.swp
