+++ src/libvmon/bitmap.h
