+++ src/libvmon/defs/_begin.def
