+++ src/libvmon/defs/proc_stat.def
