+++ src/libvmon/defs/sys_stat.def
