+++ src/logo.h
@@ -1,0 +1,8 @@
+#ifndef _LOGO_H
+#define _LOGO_H
+
+typedef struct _vwm_t vwm_t;
+
+void vwm_draw_logo(vwm_t *vwm);
+
+#endif
